+++ index.html
@@ -232,6 +232,8 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
   <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+  <script type="module" src="/js/main.js"></script>
+  <script type="module" src="/js/util.js"></script>
+  <script type="module" src="/js/data.js"></script>
 </body>
 </html>
